agents/requirements.txt
sqlalchemy>=1.4.0
python-dotenv>=1.0.0
croniter>=1.3.0
psycopg2-binary>=2.9.0
